+++ src/components/forms/PaymentOptionsForm/PaymentOptionsForm.tsx
@@ -3,6 +3,7 @@ import { USD } from '@ardrive/turbo-sdk';
 import { ArIOTokenIcon, TurboIcon } from '@src/components/icons';
 import { Checkbox } from '@src/components/inputs/Checkbox';
 import { SelectDropdown } from '@src/components/inputs/Select';
+import TurboTopUpModal from '@src/components/modals/turbo/TurboTopUpModal';
 import {
   useArIOLiquidBalance,
   useArIOStakedAndVaultedBalance,
@@ -423,6 +424,8 @@ function PaymentOptionsForm({
     return 0;
   }, [selectedCrypto, allArIOBalance, formattedARIOTicker]);
 
+  const [showTopupModal, setShowTopupModal] = useState(false);
+
   return (
     <>
       <div className="flex flex-col gap-6 w-full">
@@ -443,10 +446,24 @@ function PaymentOptionsForm({
               value="card"
               className="flex gap-3 p-3 data-[state=active]:bg-foreground rounded border border-[#222224] data-[state=active]:border-grey text-white items-center flex-1 whitespace-nowrap transition-all duration-300 disabled:opacity-50"
             >
-              <div className="flex gap-3 items-center">
-                <CreditCard className="size-5 text-grey" />
-                Card
-              </div>
+              <Tooltip
+                tooltipOverrides={{
+                  arrow: false,
+                  overlayInnerStyle: {
+                    whiteSpace: 'nowrap',
+                    width: 'fit-content',
+                    padding: '0.625rem',
+                    border: '1px solid var(--text-faded)',
+                  },
+                }}
+                message="Coming Soon!"
+                icon={
+                  <div className="flex gap-3 items-center">
+                    <CreditCard className="size-5 text-grey" />
+                    Credit Card
+                  </div>
+                }
+              />
             </Tabs.Trigger>
             <Tabs.Trigger
               value="crypto"
@@ -488,7 +505,7 @@ function PaymentOptionsForm({
               className={{
                 trigger:
                   'flex w-full gap-2 p-3 rounded-md bg-transparent border border-[#222224] items-center pointer-events-none',
-                icon: 'text-grey size-5',
+                icon: 'text-transparent size-5',
               }}
               options={cryptoDropdownOptions}
             />
@@ -513,31 +530,31 @@ function PaymentOptionsForm({
                   className="flex flex-col w-full h-full"
                 >
                   <Tabs.List
-                    className="flex flex-col w-full gap-2 text-white text-xs"
+                    className="flex flex-col w-full gap-2 text-white text-sm"
                     defaultValue={'balance'}
                   >
                     <Tabs.Trigger
                       value="balance"
-                      className="flex w-full gap-2 p-2 rounded bg-foreground data-[state=inactive]:bg-transparent border border-dark-grey items-center"
+                      className="flex w-full gap-2 p-3 rounded bg-foreground data-[state=inactive]:bg-transparent border border-dark-grey items-center"
                       onClick={() => onFundingSourceChange('balance')}
                     >
                       {fundingSource === 'balance' ? (
-                        <CircleCheck className="size-4 text-background fill-white" />
+                        <CircleCheck className="size-5 text-background fill-white" />
                       ) : (
-                        <Circle className="size-4 text-grey" />
+                        <Circle className="size-5 text-grey" />
                       )}
                       <span className="font-bold">Liquid Balance</span> (
                       {formatARIOWithCommas(liquidArIOBalance)} ARIO)
                     </Tabs.Trigger>
                     <Tabs.Trigger
                       value="any"
-                      className="flex w-full gap-2 p-2 rounded bg-foreground data-[state=inactive]:bg-transparent border border-dark-grey items-center"
+                      className="flex w-full gap-2 p-3 rounded bg-foreground data-[state=inactive]:bg-transparent border border-dark-grey items-center"
                       onClick={() => onFundingSourceChange('any')}
                     >
                       {fundingSource === 'any' ? (
-                        <CircleCheck className="size-4 text-background fill-white" />
+                        <CircleCheck className="size-5 text-background fill-white" />
                       ) : (
-                        <Circle className="size-4 text-grey" />
+                        <Circle className="size-5 text-grey" />
                       )}
                       <span className="font-bold">
                         Liquid + Staked Balances
@@ -546,13 +563,13 @@ function PaymentOptionsForm({
                     </Tabs.Trigger>
                     <Tabs.Trigger
                       value="stakes"
-                      className="flex w-full gap-2 p-2 rounded bg-foreground data-[state=inactive]:bg-transparent border border-dark-grey items-center"
+                      className="flex w-full gap-2 p-3 rounded bg-foreground data-[state=inactive]:bg-transparent border border-dark-grey items-center"
                       onClick={() => onFundingSourceChange('stakes')}
                     >
                       {fundingSource === 'stakes' ? (
-                        <CircleCheck className="size-4 text-background fill-white" />
+                        <CircleCheck className="size-5 text-background fill-white" />
                       ) : (
-                        <Circle className="size-4 text-grey" />
+                        <Circle className="size-5 text-grey" />
                       )}{' '}
                       <span className="font-bold">Staked Balances</span> (
                       {formatARIOWithCommas(stakedAndVaultedArIOBalance)} ARIO)
@@ -583,19 +600,20 @@ function PaymentOptionsForm({
                     Please top-up to complete your purchase in credits.
                   </span>
                 </div>
-                <Link
+                <button
                   className="py-2 px-6 text-lg text-white rounded bg-dark-grey border border-transparent hover:border-grey"
-                  to={`https://turbo-topup.com`}
-                  target="_blank"
-                  rel="noreferrer"
+                  onClick={() => setShowTopupModal(true)}
                 >
                   Top-Up
-                </Link>
+                </button>
               </div>
             )}
           </Tabs.Content>
         </Tabs.Root>{' '}
       </div>
+      {showTopupModal && (
+        <TurboTopUpModal onClose={() => setShowTopupModal(false)} />
+      )}
     </>
   );
 }