+++ src/components/modals/turbo/panels/crypto/CryptoConfirmation.tsx
@@ -1,0 +1,5 @@
+function CryptoConfirmation() {
+  return <div></div>;
+}
+
+export default CryptoConfirmation;